+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT8T9AETWT4CHJGPJ2C5046A"
+        "01KT8WF7C6S484M53JADTRAP82"
     )
     r"""The unique identifier of the data source"""